+++ src/test/goldens/paper-behaviors/paper-button-behavior.d.ts
@@ -42,4 +42,6 @@ declare namespace Polymer {
     _calculateElevation(): any;
     _computeKeyboardClass(receivedFocusFromKeyboard: any): any;
   }
+
+  const PaperButtonBehavior: object;
 }