+++ src/test/goldens/paper-behaviors/paper-checked-element-behavior.d.ts
@@ -31,4 +31,6 @@ declare namespace Polymer {
      */
     _checkedChanged(): any;
   }
+
+  const PaperCheckedElementBehavior: object;
 }